refactor: promote runtime read protocols
