+++ backend/chat/api/http/conversations_router.py
@@ -10,9 +10,9 @@ from fastapi import APIRouter, Depends
 
 from backend.avatar_urls import avatar_url
 from backend.chat.api.http.dependencies import get_app, get_current_user_id
-from backend.protocols.runtime_read import RuntimeThreadActivityReader
 from backend.thread_runtime.owner_reads import list_owner_thread_rows_for_auth_burst
 from backend.thread_runtime.projection import canonical_owner_threads
+from protocols.runtime_read import RuntimeThreadActivityReader
 
 router = APIRouter(prefix="/api/conversations", tags=["conversations"])
