+++ backend/protocols/runtime_read.py
@@ -1,18 +1,5 @@
-"""Runtime read-side protocol contracts."""
+"""Compatibility export surface for top-level runtime-read protocols."""
 
-from __future__ import annotations
+from protocols.runtime_read import AgentThreadActivity, RuntimeThreadActivityReader
 
-from dataclasses import dataclass
-from typing import Literal, Protocol
-
-
-@dataclass(frozen=True)
-class AgentThreadActivity:
-    thread_id: str
-    is_main: bool
-    branch_index: int
-    state: Literal["initializing", "ready", "active", "idle", "suspended", "stopped", "destroyed"]
-
-
-class RuntimeThreadActivityReader(Protocol):
-    def list_active_threads_for_agent(self, agent_user_id: str) -> list[AgentThreadActivity]: ...
+__all__ = ["AgentThreadActivity", "RuntimeThreadActivityReader"]
