+++ protocols/__init__.py
@@ -1,0 +1,1 @@
+"""Portable protocol contracts shared across non-backend package boundaries."""
