+++ tests/Unit/messaging/test_runtime_read_protocol_owner.py
@@ -1,0 +1,23 @@
+from __future__ import annotations
+
+import importlib
+from typing import get_type_hints
+
+
+def test_runtime_read_protocols_live_in_top_level_protocols_module() -> None:
+    protocol_module = importlib.import_module("protocols.runtime_read")
+    backend_protocol_module = importlib.import_module("backend.protocols.runtime_read")
+
+    assert protocol_module.AgentThreadActivity.__module__ == "protocols.runtime_read"
+    assert protocol_module.RuntimeThreadActivityReader.__module__ == "protocols.runtime_read"
+    assert backend_protocol_module.AgentThreadActivity is protocol_module.AgentThreadActivity
+    assert backend_protocol_module.RuntimeThreadActivityReader is protocol_module.RuntimeThreadActivityReader
+
+
+def test_runtime_thread_selector_consumes_top_level_runtime_read_protocol() -> None:
+    selector_module = importlib.import_module("messaging.delivery.runtime_thread_selector")
+    protocol_module = importlib.import_module("protocols.runtime_read")
+
+    hints = get_type_hints(selector_module.select_runtime_thread_for_recipient)
+
+    assert hints["activity_reader"] is protocol_module.RuntimeThreadActivityReader
